agents/pb-foundryv1-invoice-agent/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundryv1-invoice-agent"
version = "0.1.0"
description = "Foundry Agents v1 invoice assistant runtime with file search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "azure-ai-agents>=1.1.0",
  "azure-identity>=1.17.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.1",
  "pytest>=8.2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
